Changes: - Changed collisionmanager code to now be aware of which scene the collider is from, to improve performance and scenes not having collisions when not attive. - Added extra reset logic in ingredient component. - Changed level loading to now say what scene they belong to. - Added code that has sub colliders (ingredients and movement component) to now get the scene id from their parent and give it to their sub colliders.

+++ Minigin/SceneManager.cpp
@@ -58,7 +58,7 @@ void dae::SceneManager::ResetActiveScene()
 	m_pScenes[m_ActiveSceneIndex]->Reset();
 }
 
-void dae::SceneManager::SetActiveScene(size_t index)
+void dae::SceneManager::SetActiveScene(int index)
 {
 	if (index < 0 || index>m_pScenes.size() - 1)
 		return;
@@ -79,7 +79,7 @@ void dae::SceneManager::GotoPreviousScene()
 	--m_ActiveSceneIndex;
 
 	if (m_ActiveSceneIndex < 0)
-		m_ActiveSceneIndex = m_pScenes.size() - 1;
+		m_ActiveSceneIndex = static_cast<int>(m_pScenes.size()) - 1;
 }
 
 Scene& SceneManager::CreateScene(const std::string_view name)
